Move setShardAttributesToParams
Too special-purpose to warrant any back-compat concern.  Was here since 2023-06
