+++ solr/core/src/java/org/apache/solr/handler/component/TaskManagementHandler.java
@@ -79,7 +79,7 @@ public abstract class TaskManagementHandler extends RequestHandlerBase
         String reqPath = (String) req.getContext().get(PATH);
 
         params.set(CommonParams.QT, reqPath);
-        params.setShardAttributesToParams(sreq.purpose);
+        ShardHandler.setShardAttributesToParams(params, sreq.purpose);
 
         if (extraParams != null) {
           for (Map.Entry<String, String> entry : extraParams.entrySet()) {
