+++ solr/solrj/src/java/org/apache/solr/common/params/ModifiableSolrParams.java
@@ -16,9 +16,6 @@
  */
 package org.apache.solr.common.params;
 
-import static org.apache.solr.common.params.CommonParams.DISTRIB;
-import static org.apache.solr.common.params.CommonParams.INDENT;
-
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -178,16 +175,6 @@ public class ModifiableSolrParams extends SolrParams {
 
   // ----------------------------------------------------------------
   // ----------------------------------------------------------------
-
-  public void setShardAttributesToParams(int purpose) {
-    remove(ShardParams.SHARDS); // not a top-level request
-    set(DISTRIB, Boolean.FALSE.toString()); // not a top-level request
-    remove(INDENT);
-    remove(CommonParams.HEADER_ECHO_PARAMS);
-    set(ShardParams.IS_SHARD, true); // a sub (shard) request
-    set(ShardParams.SHARDS_PURPOSE, purpose);
-    set(CommonParams.OMIT_HEADER, false);
-  }
 
   @Override
   public String get(String param) {
